+++ apps/web/src/app/[locale]/layout.tsx
@@ -12,12 +12,12 @@ import type React from "react";
 import { TimeZoneChangeDetector } from "@/app/[locale]/timezone-change-detector";
 import { UserProvider } from "@/components/user-provider";
 import { PreferencesProvider } from "@/contexts/preferences";
-import { FeatureFlagsProvider } from "@/features/feature-flags/client";
 import { isStorageEnabled } from "@/features/storage";
 import { TimezoneProvider } from "@/features/timezone/client/context";
 import { getUser } from "@/features/user/queries";
 import { I18nProvider } from "@/i18n/client";
 import { getLocale } from "@/i18n/server/get-locale";
+import { FeatureFlagsProvider } from "@/lib/feature-flags/client";
 import { auth } from "@/next-auth";
 import { TRPCProvider } from "@/trpc/client/provider";
 import { ConnectedDayjsProvider } from "@/utils/dayjs";
